requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.34.0
sqlalchemy[asyncio]==2.0.36
aiosqlite==0.20.0
python-dotenv==1.0.1
python-jose[cryptography]==3.3.0
pydantic[email]==2.10.4
jinja2==3.1.5
python-multipart==0.0.18
httpx==0.28.1
aiosmtplib==3.0.2
anthropic==0.43.0
apscheduler==3.10.4
